# backend/docs/swagger/swagger.yaml
basePath: /
definitions:
  ApiKeyCreateRequest:
    properties:
      name:
        example: My API Key
        maxLength: 50
        minLength: 1
        type: string
      user_id:
        example: 01234567-89ab-cdef-0123-456789abcdef
        format: uuid
        type: string
    required:
    - name
    - user_id
    type: object
  ApiKeyItem:
    properties:
      created_at:
        type: string
      id:
        example: 01234567-89ab-cdef-0123-456789abcdef
        type: string
      key:
        example: todo_abc123...
        type: string
      name:
        example: My API Key
        type: string
      user_id:
        example: 01234567-89ab-cdef-0123-456789abcdef
        type: string
    type: object
  ErrorResponse:
    properties:
      message:
        example: Not Found
        type: string
    type: object
  HealthCheckResponse:
    properties:
      status:
        example: ok
        type: string
    type: object
  TodoItem:
    properties:
      created_at:
        type: string
      description:
        example: 牛乳、卵、パンを買う
        type: string
      due_date:
        type: string
      id:
        example: 01234567-89ab-cdef-0123-456789abcdef
        type: string
      sort_order:
        type: number
      status:
        example: pending
        type: string
      tags:
        items:
          $ref: '#/definitions/internal_presentation_http_todo.todoTag'
        type: array
      title:
        example: 買い物リストを作る
        type: string
      updated_at:
        type: string
    type: object
  TodoUpdateRequest:
    properties:
      clear_due_date:
        type: boolean
      description:
        example: 牛乳、卵、パンを買う
        maxLength: 1000
        type: string
      due_date:
        type: string
      sort_order:
        type: number
      status:
        enum:
        - pending
        - in_progress
        - completed
        example: in_progress
        type: string
      title:
        example: 買い物リストを作る
        maxLength: 100
        type: string
    type: object
  UserCreateRequest:
    properties:
      email:
        example: john@example.com
        format: email
        type: string
      name:
        example: John Doe
        maxLength: 100
        minLength: 1
        type: string
    required:
    - email
    - name
    type: object
  UserItem:
    properties:
      created_at:
        type: string
      email:
        example: john@example.com
        type: string
      id:
        example: 01234567-89ab-cdef-0123-456789abcdef
        type: string
      name:
        example: John Doe
        type: string
      updated_at:
        type: string
    type: object
  internal_presentation_http_tag.createRequest:
    properties:
      name:
        type: string
    required:
    - name
    type: object
  internal_presentation_http_tag.setTodoTagsRequest:
    properties:
      tag_ids:
        items:
          type: string
        type: array
    type: object
  internal_presentation_http_tag.tagItem:
    properties:
      created_at:
        type: string
      id:
        type: string
      name:
        type: string
      updated_at:
        type: string
    type: object
  internal_presentation_http_todo.todoTag:
    properties:
      id:
        type: string
      name:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: Todo管理アプリケーションのREST API
  title: Todo API
  version: "1.0"
paths:
  /health:
    get:
      description: サーバーの稼働状況を確認します
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/HealthCheckResponse'
      summary: ヘルスチェック
      tags:
      - health
  /v1/api-keys:
    post:
      consumes:
      - application/json
      description: 新しいAPIキーを発行します（認証不要）
      parameters:
      - description: APIキー作成リクエスト
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/ApiKeyCreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/ApiKeyItem'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: APIキー作成
      tags:
      - api-keys
  /v1/api-keys/{id}:
    delete:
      description: 指定されたIDのAPIキーを削除します
      parameters:
      - description: APIキーID
        example: '"01234567-89ab-cdef-0123-456789abcdef"'
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - ApiKeyAuth: []
      - SkipAuth: []
      summary: APIキー削除
      tags:
      - api-keys
  /v1/tags:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/internal_presentation_http_tag.tagItem'
            type: array
      security:
      - ApiKeyAuth: []
      - SkipAuth: []
      summary: タグ一覧取得
      tags:
      - tags
    post:
      consumes:
      - application/json
      parameters:
      - description: タグ作成リクエスト
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_presentation_http_tag.createRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/internal_presentation_http_tag.tagItem'
      security:
      - ApiKeyAuth: []
      - SkipAuth: []
      summary: タグ作成
      tags:
      - tags
  /v1/tags/{id}:
    delete:
      parameters:
      - description: タグID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
      security:
      - ApiKeyAuth: []
      - SkipAuth: []
      summary: タグ削除
      tags:
      - tags
  /v1/todos:
    get:
      description: すべてのタスクを一覧で取得します
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/TodoItem'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - ApiKeyAuth: []
      - SkipAuth: []
      summary: タスク一覧取得
      tags:
      - todos
  /v1/todos/{id}:
    delete:
      description: 指定されたIDのタスクを削除します
      parameters:
      - description: タスクID
        example: '"01234567-89ab-cdef-0123-456789abcdef"'
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - ApiKeyAuth: []
      - SkipAuth: []
      summary: タスク削除
      tags:
      - todos
    get:
      description: 指定されたIDのタスクを取得します
      parameters:
      - description: タスクID
        example: '"01234567-89ab-cdef-0123-456789abcdef"'
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/TodoItem'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - ApiKeyAuth: []
      - SkipAuth: []
      summary: タスク詳細取得
      tags:
      - todos
    put:
      consumes:
      - application/json
      description: 指定されたIDのタスクを更新します
      parameters:
      - description: タスクID
        example: '"01234567-89ab-cdef-0123-456789abcdef"'
        in: path
        name: id
        required: true
        type: string
      - description: タスク更新リクエスト
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/TodoUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/TodoItem'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - ApiKeyAuth: []
      - SkipAuth: []
      summary: タスク更新
      tags:
      - todos
  /v1/todos/{id}/tags:
    put:
      consumes:
      - application/json
      parameters:
      - description: TodoID
        in: path
        name: id
        required: true
        type: string
      - description: タグIDリスト
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_presentation_http_tag.setTodoTagsRequest'
      responses:
        "204":
          description: No Content
      security:
      - ApiKeyAuth: []
      - SkipAuth: []
      summary: TodoにタグをセットするPUT /v1/todos/:id/tags
      tags:
      - todos
  /v1/users:
    post:
      consumes:
      - application/json
      description: 新しいユーザーを作成します
      parameters:
      - description: ユーザー作成リクエスト
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/UserCreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/UserItem'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - ApiKeyAuth: []
      - SkipAuth: []
      summary: ユーザー作成
      tags:
      - users
securityDefinitions:
  ApiKeyAuth:
    description: X-API-Keyヘッダーにトークンを設定してください
    in: header
    name: X-API-Key
    type: apiKey
  SkipAuth:
    description: 認証をスキップするためのヘッダー。テストや特定の内部APIで使用する。ユーザーIDを指定する。
    in: header
    name: X-Skip-Auth
    type: apiKey
swagger: "2.0"