# backend/internal/presentation/http/todo/delete.go
package todo

import (
	"net/http"

	"github.com/gin-gonic/gin"
	"github.com/jimi024rion/todo-go/backend/internal/config/errs"
	"github.com/jimi024rion/todo-go/backend/internal/presentation/http/response"
	todousecase "github.com/jimi024rion/todo-go/backend/internal/usecase/todo"
)

type DeleteHandler struct {
	usecase *todousecase.DeleteUseCase
}

func NewDeleteHandler(u *todousecase.DeleteUseCase) *DeleteHandler {
	return &DeleteHandler{usecase: u}
}

// Handle handles the request to delete a todo.
//
// @Summary     タスク削除
// @Description 指定されたIDのタスクを削除します
// @Tags        todos
// @Produce     json
// @Param       id  path     string true "タスクID" example("01234567-89ab-cdef-0123-456789abcdef")
// @Success     204
// @Failure     400 {object} response.ErrorResponse
// @Failure     404 {object} response.ErrorResponse
// @Failure     500 {object} response.ErrorResponse
// @Security    ApiKeyAuth
// @Security    SkipAuth
// @Router      /v1/todos/{id} [delete]
func (h *DeleteHandler) Handle(c *gin.Context) {
	id := c.Param("id")

	_, err := h.usecase.Execute(c.Request.Context(), &todousecase.DeleteInput{ID: id})
	if err != nil {
		rc := errs.ResultCodeFrom(err)
		c.JSON(rc.HTTPStatus(), response.Fail(rc.Message()))
		return
	}

	c.Status(http.StatusNoContent)
}